refactor: Extract Feed and Profile into seprate components
