+++ app/(tabs)/index.tsx
@@ -1,23 +1,6 @@
-import {
-  Actions,
-  ActionsSkeleton,
-  Activity,
-  ActivitySkeleton,
-  Badges,
-  BadgesSkeleton,
-  PointsView,
-  PointsViewSkeleton,
-  ProfileHeader,
-  ProfileHeaderSkeleton,
-} from "@/lib/components/Profile";
-import { ThemedView } from "@/lib/components/ThemedView";
 import { Colors } from "@/lib/constants/Colors";
-import { useActivityFeed } from "@/lib/net/queries/post/useActivityFeed";
-import React, { PropsWithChildren, useState, useEffect } from "react";
+import React, { PropsWithChildren, useState } from "react";
 import {
-  ActivityIndicator,
-  FlatList,
-  ScrollView,
   StyleSheet,
   useWindowDimensions,
   View,
@@ -29,16 +12,9 @@ import {
   TabBarIndicator,
   TabView,
 } from "react-native-tab-view";
-import { EmptyViewWithTabBarHeight } from "@/lib/components/util";
-import { Post } from "@/lib/types";
-import { ThemedText } from "@/lib/components/ThemedText";
-import { useThemeColor } from "@/lib/hooks/useThemeColor";
-import { ProposalCard, ProposalCardSkeleton } from "@/lib/components/proposal/ProposalCard";
-import { useProfileStore } from "@/lib/store/profileStore";
-import { SafeAreaView, useSafeAreaInsets } from "react-native-safe-area-context";
-import { useGetUserById } from "@/lib/net/queries/profile";
-import getIdFromToken from "@/lib/util/jwt";
-import { useAuthStore } from "@/lib/store/authStore";
+import { SafeAreaView } from "react-native-safe-area-context";
+import { Profile } from "@/lib/components/Profile";
+import { Feed } from "@/lib/components/feed";
 
 const renderScene = SceneMap({
   profile: Profile,
@@ -49,107 +25,6 @@ const routes = [
   { key: "profile", title: "Profile" },
   { key: "feed", title: "Feed" },
 ];
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1
-  },
-  scrollView: {
-    paddingHorizontal: 16,
-    marginTop: 16,
-  },
-});
-
-function Profile() {
-  const userProfile = useProfileStore((state) => state.profile);
-  const accessToken = useAuthStore((state) => state.accessToken);
-
-  const userId = getIdFromToken(accessToken || "");
-  const { data } = useGetUserById(userId || "");
-
-  useEffect(() => {
-    if (data) {
-      useProfileStore.setState({ profile: data });
-    }
-  }, [data]);
-
-  if (!userProfile) {
-    return <ProfileSkeleton />;
-  }
-
-  return (
-    <ThemedView type="background" style={styles.container}>
-      <ScrollView style={styles.scrollView} contentContainerStyle={{ gap: 20 }}>
-        <ProfileHeader username={userProfile.username} avatarUrl={userProfile.profileDetails?.image} />
-        <PointsView points={userProfile.profileScore} />
-        <Badges />
-        <Actions />
-        <Activity userId={String(userProfile.id)} />
-        <EmptyViewWithTabBarHeight />
-      </ScrollView>
-    </ThemedView>
-  );
-}
-
-const ProfileSkeleton = () => (
-  <ThemedView type="background" style={{ gap: 20, marginTop: 16, paddingHorizontal: 16 }}>
-    <ProfileHeaderSkeleton />
-    <PointsViewSkeleton />
-    <BadgesSkeleton />
-    <ActionsSkeleton />
-    <ActivitySkeleton />
-  </ThemedView>
-);
-
-function Feed() {
-  const {
-    data,
-    fetchNextPage,
-    hasNextPage,
-    isFetchingNextPage,
-    isLoading,
-  } = useActivityFeed({ limit: 10 });
-
-  const renderItem = ({ item }: { item: Post }) => <ProposalCard post={item} />;
-  const accentColor = useThemeColor({}, "accent");
-
-  if (isLoading || !data) {
-    return (
-      <ThemedView type="background" style={[styles.container, { justifyContent: "center" }]}>
-
-        <ProposalCardSkeleton />
-        <ProposalCardSkeleton />
-        <ProposalCardSkeleton />
-      </ThemedView>
-    );
-  }
-
-  return (
-    <ThemedView type="background" style={styles.container}>
-      <FlatList
-        contentContainerStyle={{
-          gap: 8,
-          paddingInline: 8,
-        }}
-        data={data?.pages.flatMap((page) => page.items)}
-        renderItem={renderItem}
-        keyExtractor={(item) => item.index.toString()}
-        onEndReached={() => {
-          if (hasNextPage && !isFetchingNextPage) {
-            fetchNextPage();
-          }
-        }}
-        onEndReachedThreshold={0.5}
-        ListFooterComponent={
-          <EmptyViewWithTabBarHeight>
-            {isFetchingNextPage ? <ActivityIndicator size="small" color={accentColor} style={{ marginBottom: 20 }} /> : <ThemedText></ThemedText>}
-          </EmptyViewWithTabBarHeight>
-        }
-        ListEmptyComponent={<EmptyViewWithTabBarHeight />}
-      />
-    </ThemedView>
-  );
-}
 
 export default function Home() {
   const layout = useWindowDimensions();
@@ -166,7 +41,7 @@ export default function Home() {
               {...props}
               style={{
                 height: 52,
-                backgroundColor: "#00000000",
+                backgroundColor: "transparent",
               }}
               renderIndicator={(props) => (
                 <TabBarIndicator
