+++ lib/components/Profile/index.ts
@@ -1,5 +1,3 @@
-export { ProfileHeader, ProfileHeaderSkeleton } from "./ProfileHeader";
-export { PointsView, PointsViewSkeleton } from "./PointsView";
-export { Badges, BadgesSkeleton } from "./Badges";
-export { Actions, ActionsSkeleton } from "./Actions";
-export { Activity, ActivitySkeleton } from "./Activity";
+import { Profile } from "./Profile";
+
+export { Profile };
