+++ lib/components/feed/index.ts
@@ -1,5 +1,6 @@
 import CommentBox from "./CommentBox";
 import CommentCard from "./CommentCard";
 import CommentList from "./CommentList";
+import Feed from "./Feed";
 
-export { CommentBox, CommentCard, CommentList } 
+export { CommentBox, CommentCard, CommentList, Feed }
